% Auto-generated: do not edit by hand
\name{webSocket}

\alias{webSocket}

\title{WebSocket component}

\description{

}

\usage{
webSocket(id=NULL, autoReconnect=NULL, close=NULL, error=NULL,
kill=NULL, kill_and_restart=NULL, maxReconnectAttempts=NULL,
message=NULL, pingInterval=NULL, protocols=NULL,
reconnect=NULL, reconnectInterval=NULL, restart=NULL,
send=NULL, state=NULL, url=NULL)
}

\arguments{
\item{id}{Character. }

\item{autoReconnect}{Logical. }

\item{close}{Named list | character. }

\item{error}{Named list | character. }

\item{kill}{Named list | character. }

\item{kill_and_restart}{Named list | character. }

\item{maxReconnectAttempts}{Numeric. }

\item{message}{Named list | character. }

\item{pingInterval}{Numeric. }

\item{protocols}{List of characters. }

\item{reconnect}{Named list | character. }

\item{reconnectInterval}{Numeric. }

\item{restart}{Named list | character. }

\item{send}{Named list | character. }

\item{state}{Named list | character. }

\item{url}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

